moved like fetch to app component, restructured code
